Update search UI
The suggestion list hides when it, and the other search controls (input field and button) are not in focus. You can navigate the list with the arrow keys. Decreased the amount of suggestions to 10 to better fit. Replaced the "No suggestions" text with a link to searching for the term on WP.

Bug: T323926