Add banner to the top of the page
Banner content can be loaded via endpoint /set-banner from the URL specified in the config. If specified, a CSS selector is used to get content from a specific element.

Endpoint /preview-banner can be used to preview a banner without saving it. URL and selector can be specified through URL parameters.

Config is loaded the first time for a request and then reused.

Bug: T329613